Update cisco.cson
I fixed the banner-motd of the day search, instead of searching until the end of the file after the motd, it now searches until it encounters the ^ character again. I then duplicated it for multiple banners that I've come across. Currently includes exec, login, and prompt-timeout. Tested and working. -MB